apps/agent-service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
